[Crashlytics] AE-37 Fix NPE when accessing billing manager
Crash was happening when the app is opened and then exited using back and then opened after a long time. This long time led the OS to kill the application process leading to fragments being recreated. Fragments had
dependencies injected from the Activity. And since fragment recreation
callback was occuring before the Activity onCreate, the variables were null.
(Use Don't keep activities in Dev options to recreate the scenario without waiting)

Fix was to use interface as mode of communication btw fragment and Activity
as per Android Developer guidelines. Also, to not create fragments again
if savedInstance already exists.

In future, a shared ViewModel can be used to fix this issue.