Fix bug in Recent date calculation
Recent was using date in ms to fetch all items within x days from
present date. But this logic fails when the date is within x days
but the time is different leading to some last day files before
the time ms to not show.
Approach was to use calendar to find the files within x days.